feat(core): added aliases to all functions in index files * moved a lot of utilities to the relevant new folders (measurements, validation etc)
