+++ src/core/geometry/shape2/fromCompactBinary.js
@@ -31,4 +31,4 @@ const fromCompactBinary = function (bin) {
   return cag
 }
 
-module.exports = fromSides
+module.exports = fromCompactBinary
