+++ src/core/geometry/shape2/index.js
@@ -1,0 +1,30 @@
+module.exports = {
+  canonicalize: require('./canonicalize'),
+  create: require('./create'),
+  fromCompactBinary: require('./fromCompactBinary'),
+  fromFakeCSG: require('./fromFakeCSG'),
+  fromObject: require('./fromObject'),
+  fromPoints: require('./fromPoints'),
+  fromPointsNoCheck: require('./fromPointsNoCheck'),
+  fromSides: require('./fromSides'),
+  equals: require('./equals'),
+  measureArea: require('./measureArea'),
+  measureBounds: require('./measureBounds'),
+  mirror: require('./mirror'),
+  mirrorX: require('./mirrorX'),
+  mirrorY: require('./mirrorY'),
+  mirrorZ: require('./mirrorZ'),
+  rotate: require('./rotate'),
+  rotateX: require('./rotateX'),
+  rotateY: require('./rotateY'),
+  rotateZ: require('./rotateZ'),
+  scale: require('./scale'),
+  toCompactBinary: require('./toCompactBinary'),
+  toCSGWall: require('./toCSGWall'),
+  toPoints: require('./toPoints'),
+  toString: require('./toString'),
+  subtract: require('./subtract'),
+  transform: require('./transform'),
+  validate: require('./validate'),
+  validateIsSelfIntersecting: require('./validateIsSelfIntersecting')
+}
