+++ src/core/geometry/shape3/csgFeatures.js
