+++ src/core/geometry/shape3/index.js
@@ -1,0 +1,32 @@
+module.exports = {
+  canonicalize: require('./canonicalize'),
+  connectTo: require('./connectTo'),
+  create: require('./create'),
+  equals: require('./equals'),
+  fromCompactBinary: require('./fromCompactBinary'),
+  fromObject: require('./fromObject'),
+  fromPolygons: require('./fromPolygons'),
+  fromSlices: require('./fromSlices'),
+  invert: require('./invert'),
+  measureArea: require('./measureArea'),
+  measureBounds: require('./measureBounds'),
+  measureVolume: require('./measureVolume'),
+  mirror: require('./mirror'),
+  mirrorX: require('./mirrorX'),
+  mirrorY: require('./mirrorY'),
+  mirrorZ: require('./mirrorZ'),
+  rotate: require('./rotate'),
+  rotateX: require('./rotateX'),
+  rotateY: require('./rotateY'),
+  rotateZ: require('./rotateZ'),
+  scale: require('./scale'),
+  setColor: require('./setColor'),
+  setShared: require('./setShared'),
+  projectToOrthoNormalBasis: require('./projectToOrthoNormalBasis'),
+  toCompactBinary: require('./toCompactBinary'),
+  toString: require('./toString'),
+  toTriangles: require('./toTriangles'),
+  subtract: require('./subtract'),
+  transform: require('./transform'),
+  translate: require('./translate'),
+}
