+++ src/core/geometry/shape3/measureArea.js
@@ -1,0 +1,8 @@
+const area = function (csg) {
+  let result = csg.toTriangles().map(function (triPoly) {
+    return triPoly.getTetraFeatures(['area'])
+  })
+  console.log('area', result)
+}
+
+module.exports = area
