+++ src/core/geometry/shape3/measureBounds.js
@@ -1,13 +1,13 @@
 const Vector3D = require('../math/Vector3')
 
 /**
-   * Returns an array of Vector3D, providing minimum coordinates and maximum coordinates
-   * of this solid.
-   * @returns {Vector3D[]}
-   * @example
-   * let bounds = A.getBounds()
-   * let minX = bounds[0].x
-   */
+ * Returns an array of Vector3D, providing minimum coordinates and maximum coordinates
+ * of this solid.
+ * @returns {Vector3D[]}
+ * @example
+ * let bounds = A.getBounds()
+ * let minX = bounds[0].x
+ */
 const bounds = function (csg) {
   if (!csg.cachedBoundingBox) {
     let minpoint = new Vector3D(0, 0, 0)
@@ -31,18 +31,4 @@ const bounds = function (csg) {
   return csg.cachedBoundingBox
 }
 
-const volume = function (csg) {
-  let result = csg.toTriangles().map(function (triPoly) {
-    return triPoly.getTetraFeatures(['volume'])
-  })
-  console.log('volume', result)
-}
-
-const area = function (csg) {
-  let result = csg.toTriangles().map(function (triPoly) {
-    return triPoly.getTetraFeatures(['area'])
-  })
-  console.log('area', result)
-}
-
-module.exports = {bounds, volume, area}
+module.exports = bounds
