+++ src/core/geometry/shape3/measureVolume.js
@@ -1,0 +1,8 @@
+const volume = function (csg) {
+  let result = csg.toTriangles().map(function (triPoly) {
+    return triPoly.getTetraFeatures(['volume'])
+  })
+  console.log('volume', result)
+}
+
+module.exports = volume
