+++ src/core/geometry/shape3/shape3.test.js
@@ -1,0 +1,8 @@
+const test = require('ava')
+const create = require('./create')
+
+test('New shape3 should contain nothing', t => {
+  const shape3 = create()
+  console.log('shape3', shape3)
+
+})
