+++ src/core/math/vec2/index.js
@@ -1,0 +1,33 @@
+module.exports = {
+  abs: require('./abs'),
+  add: require('./add'),
+  angle: require('./angle'),
+  angleDegree: require('./angleDegrees'),
+  angleRadians: require('./angleRadians'),
+  clone: require('./clone'),
+  create: require('./create'),
+  cross: require('./cross'),
+  distance: require('./distance'),
+  divide: require('./divide'),
+  dot: require('./dot'),
+  equals: require('./equals'),
+  fromAngle: require('./fromAngle'),
+  fromAngleDegrees: require('./fromAngleDegrees'),
+  fromAngleRadians: require('./fromAngleRadians'),
+  fromValues: require('./fromValues'),
+  fromVarious: require('./fromVarious'),
+  length: require('./length'),
+  lerp: require('./lerp'),
+  max: require('./max'),
+  min: require('./min'),
+  multiply: require('./multiply'),
+  negate: require('./negate'),
+  normal: require('./normal'),
+  normalize: require('./normalize'),
+  rotate: require('./rotate'),
+  squaredDistance: require('./squaredDistance'),
+  squaredLength: require('./squaredLength'),
+  subtract: require('./subtract'),
+  toString: require('./toString'),
+  transformMat4: require('./transformMat4')
+}
