+++ src/core/math/vec3/index.js
@@ -1,0 +1,35 @@
+module.exports = {
+  abs: require('./abs'),
+  add: require('./add'),
+  angle: require('./angle'),
+  clone: require('./clone'),
+  create: require('./create'),
+  cross: require('./cross'),
+  distance: require('./distance'),
+  divide: require('./divide'),
+  dot: require('./dot'),
+  equals: require('./equals'),
+  fromValues: require('./fromValues'),
+  fromVarious: require('./fromVarious'),
+  fromVec2: require('./fromVec2'),
+  length: require('./length'),
+  lerp: require('./lerp'),
+  max: require('./max'),
+  min: require('./min'),
+  multiply: require('./multiply'),
+  negate: require('./negate'),
+  normal: require('./normal'),
+  normalize: require('./normalize'),
+  randomNonParallel: require('./randomNonParallel'),
+  rotate: require('./rotate'),
+  rotateX: require('./rotateX'),
+  rotateY: require('./rotateY'),
+  rotateZ: require('./rotateZ'),
+  scale: require('./scale'),
+  squaredDistance: require('./squaredDistance'),
+  squaredLength: require('./squaredLength'),
+  subtract: require('./subtract'),
+  toString: require('./toString'),
+  transform: require('./transform'),
+  transformMat4: require('./transformMat4')
+}
